+ IS(2) 20220904 - refactoring part 2
