+++ src/Account/Woden.Account.Application/Auth/Interfaces/IJwtTokenGenerator.cs
@@ -1,0 +1,5 @@
+﻿namespace KgNet88.Woden.Account.Application.Auth.Interfaces;
+public interface IJwtTokenGenerator
+{
+    public string GenerateToken(Guid userId, string username, string email);
+}
