+++ src/Account/Woden.Account.Application/GlobalUsings.cs
@@ -1,5 +1,8 @@
-﻿global using KgNet88.Woden.Account.Application.Interfaces.Auth;
+﻿global using KgNet88.Woden.Account.Application.Auth.Interfaces;
+global using KgNet88.Woden.Account.Application.Auth.Persistence;
+global using KgNet88.Woden.Account.Application.Auth.Services;
 global using KgNet88.Woden.Account.Application.Services;
+global using KgNet88.Woden.Account.Domain.Entities.Auth;
 
 global using Microsoft.Extensions.DependencyInjection;
