--- src/Account/Woden.Account.Application/Interfaces/Auth/IJwtTokenGenerator.cs
@@ -1,5 +1,0 @@
-﻿namespace KgNet88.Woden.Account.Application.Interfaces.Auth;
-public interface IJwtTokenGenerator
-{
-    public string GenerateToken(Guid userId, string username, string email);
-}
