--- src/Account/Woden.Account.Application/Interfaces/Auth/UserDto.cs
@@ -1,8 +1,0 @@
-﻿namespace KgNet88.Woden.Account.Application.Interfaces.Auth;
-
-public record UserDto
-{
-    public required Guid Id { get; init; }
-    public required string Username { get; init; }
-    public required string Email { get; init; }
-}
