--- src/Account/Woden.Account.Domain/Auth/User.cs
@@ -1,7 +1,0 @@
-﻿namespace KgNet88.Woden.Account.Domain.Auth;
-
-public class User
-{
-    public required string Username { get; init; }
-    public required string Email { get; init; }
-}
