+++ src/Account/Woden.Account.Domain/Entities/Auth/User.cs
@@ -1,0 +1,8 @@
+﻿namespace KgNet88.Woden.Account.Domain.Entities.Auth;
+
+public class User
+{
+    public required Guid Id { get; init; }
+    public required string Username { get; init; }
+    public required string Email { get; init; }
+}
