--- src/Account/Woden.Account.Infrastructure/Auth/Database/Model/DbRoleClaim.cs
@@ -1,6 +1,0 @@
-﻿namespace KgNet88.Woden.Account.Infrastructure.Auth.Database.Model;
-
-public class DbRoleClaim : IdentityRoleClaim<Guid>
-{
-    public virtual DbRole? Role { get; init; }
-}
