--- src/Account/Woden.Account.Infrastructure/Auth/Database/Model/DbUser.cs
@@ -1,8 +1,0 @@
-﻿namespace KgNet88.Woden.Account.Infrastructure.Auth.Database.Model;
-public class DbUser : IdentityUser<Guid>
-{
-    public virtual ICollection<DbUserClaim>? Claims { get; init; }
-    public virtual ICollection<DbUserLogin>? Logins { get; init; }
-    public virtual ICollection<DbUserToken>? Tokens { get; init; }
-    public virtual ICollection<DbUserRole>? UserRoles { get; init; }
-}
