--- src/Account/Woden.Account.Infrastructure/Auth/Database/Model/DbUserClaim.cs
@@ -1,6 +1,0 @@
-﻿namespace KgNet88.Woden.Account.Infrastructure.Auth.Database.Model;
-
-public class DbUserClaim : IdentityUserClaim<Guid>
-{
-    public virtual DbUser? User { get; init; }
-}
