--- src/Account/Woden.Account.Infrastructure/Auth/Database/Model/DbUserLogin.cs
@@ -1,6 +1,0 @@
-﻿namespace KgNet88.Woden.Account.Infrastructure.Auth.Database.Model;
-
-public class DbUserLogin : IdentityUserLogin<Guid>
-{
-    public virtual DbUser? User { get; init; }
-}
