+++ src/Account/Woden.Account.Infrastructure/Auth/Persistence/Database/Model/DbUser.cs
@@ -1,0 +1,8 @@
+﻿namespace KgNet88.Woden.Account.Infrastructure.Auth.Persistence.Database.Model;
+public class DbUser : IdentityUser<Guid>
+{
+    public virtual ICollection<DbUserClaim>? Claims { get; init; }
+    public virtual ICollection<DbUserLogin>? Logins { get; init; }
+    public virtual ICollection<DbUserToken>? Tokens { get; init; }
+    public virtual ICollection<DbUserRole>? UserRoles { get; init; }
+}
