+++ src/Account/Woden.Account.Infrastructure/Auth/Persistence/Database/Model/DbUserLogin.cs
@@ -1,0 +1,6 @@
+﻿namespace KgNet88.Woden.Account.Infrastructure.Auth.Persistence.Database.Model;
+
+public class DbUserLogin : IdentityUserLogin<Guid>
+{
+    public virtual DbUser? User { get; init; }
+}
