+++ src/Account/Woden.Account.Infrastructure/GlobalUsings.cs
@@ -6,19 +6,22 @@ global using FastEndpoints.Security;
 global using FluentValidation;
 global using FluentValidation.Results;
 
-global using KgNet88.Woden.Account.Application.Interfaces.Auth;
-global using KgNet88.Woden.Account.Application.Interfaces.Common;
-global using KgNet88.Woden.Account.Infrastructure.Auth;
-global using KgNet88.Woden.Account.Infrastructure.Auth.Database;
-global using KgNet88.Woden.Account.Infrastructure.Auth.Database.Model;
-global using KgNet88.Woden.Account.Infrastructure.Auth.Repositories;
-global using KgNet88.Woden.Account.Infrastructure.Common;
+global using KgNet88.Woden.Account.Application.Auth.Interfaces;
+global using KgNet88.Woden.Account.Application.Auth.Persistence;
+global using KgNet88.Woden.Account.Application.Common.Interfaces;
+global using KgNet88.Woden.Account.Domain.Entities.Auth;
+global using KgNet88.Woden.Account.Infrastructure.Auth.Implementations;
+global using KgNet88.Woden.Account.Infrastructure.Auth.Persistence;
+global using KgNet88.Woden.Account.Infrastructure.Auth.Persistence.Database;
+global using KgNet88.Woden.Account.Infrastructure.Auth.Persistence.Database.Model;
+global using KgNet88.Woden.Account.Infrastructure.Common.Implementations;
 
 global using Microsoft.AspNetCore.Identity;
 global using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 global using Microsoft.EntityFrameworkCore;
 global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.Options;
 
 global using NodaTime;
 global using NodaTime.Extensions;
